Clean up Zip4J engine and packages
