+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/ComputerDisplayZip.kt
@@ -4,6 +4,7 @@ import org.worldcubeassociation.tnoodle.server.webscrambles.pdf.ScrambleSheet
 import org.worldcubeassociation.tnoodle.server.webscrambles.pdf.engine.IText5Engine
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.util.StringUtil.randomPasscode
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.model.ZipArchive
+import org.worldcubeassociation.tnoodle.server.webscrambles.zip.model.dsl.zipArchive
 
 data class ComputerDisplayZip(val scrambleSets: Map<String, ScrambleSheet>, val competitionTitle: String) {
     val passcodes = scrambleSets.mapValues { randomPasscode() }
