+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/InterchangeFolder.kt
@@ -1,4 +1,4 @@
-package org.worldcubeassociation.tnoodle.server.webscrambles.zip.folder
+package org.worldcubeassociation.tnoodle.server.webscrambles.zip
 
 import org.worldcubeassociation.tnoodle.server.serial.JsonConfig
 import org.worldcubeassociation.tnoodle.server.webscrambles.pdf.ScrambleSheet
@@ -6,8 +6,8 @@ import org.worldcubeassociation.tnoodle.server.webscrambles.zip.util.StringUtil.
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.util.StringUtil.stripNewlines
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Competition
 import org.worldcubeassociation.tnoodle.server.webscrambles.serial.ZipInterchangeInfo
-import org.worldcubeassociation.tnoodle.server.webscrambles.zip.folder
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.model.Folder
+import org.worldcubeassociation.tnoodle.server.webscrambles.zip.model.dsl.folder
 import java.time.LocalDateTime
 
 data class InterchangeFolder(val wcif: Competition, val uniqueTitles: Map<String, ScrambleSheet>, val globalTitle: String) {
@@ -29,7 +29,7 @@ data class InterchangeFolder(val wcif: Competition, val uniqueTitles: Map<String
             folder("txt") {
                 for ((uniqueTitle, scrambleSheet) in uniqueTitles) {
                     val scrambleLines = scrambleSheet.scrambles.flatMap { it.allScrambleStrings }
-                    val txtScrambles = scrambleLines.stripNewlines().joinToString("\r\n")
+                    val txtScrambles = scrambleLines.joinToString("\r\n") { it.stripNewlines() }
                     file("$uniqueTitle.txt", txtScrambles)
                 }
             }
