+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/util/StringUtil.kt
@@ -1,5 +1,6 @@
 package org.worldcubeassociation.tnoodle.server.webscrambles.zip.util
 
+import org.apache.commons.lang3.StringUtils
 import java.security.SecureRandom
 
 object StringUtil {
@@ -10,7 +11,8 @@ object StringUtil {
     private const val PASSCODE_NUM_CHARS = 8
 
     fun String.toFileSafeString() = filter { it !in INVALID_CHARS }
-    fun List<String>.stripNewlines() = map { it.replace("\n", " ") }
+    fun String.stripNewlines() = lines().joinToString(" ")
+    fun String.stripDiacritics(): String = StringUtils.stripAccents(this)
 
     fun randomPasscode(): String {
         val secureRandom = SecureRandom()
@@ -19,5 +21,25 @@ object StringUtil {
             .map { secureRandom.nextInt(PASSCODE_DIGIT_SET.length) }
             .map { PASSCODE_DIGIT_SET[it] }
             .joinToString("")
+    }
+
+    private tailrec fun String.toUniqueTitle(seenTitles: Set<String>, suffixSalt: Int = 0): String {
+        val suffixedTitle = "$this (${suffixSalt})"
+            .takeUnless { suffixSalt == 0 } ?: this
+
+        if (suffixedTitle !in seenTitles) {
+            return suffixedTitle
+        }
+
+        return toUniqueTitle(seenTitles, suffixSalt + 1)
+    }
+
+    fun <T> List<T>.withUniqueTitles(titleGen: (T) -> String = { it.toString() }): Map<String, T> {
+        return fold(emptyMap()) { acc, req ->
+            val fileTitle = titleGen(req).toFileSafeString()
+            val safeTitle = fileTitle.toUniqueTitle(acc.keys)
+
+            acc + (safeTitle to req)
+        }
     }
 }
